requirements.txt
rich
typer
cryptography
jinja2
fpdf2
requests
aiohttp
stem